% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cube}
\alias{cube}
\title{Cube a vector of values}
\usage{
cube(x = 1:5)
}
\arguments{
\item{x}{vector of values.}
}
\description{
Simple test function that demonstrates some of the features of
  this package by cubing an input vector of values.
}
\examples{
# Find cube of first 100 values
cube(1:100)
}
