% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runMCMC}
\alias{runMCMC}
\title{Run the HMMERTIME model}
\usage{
runMCMC(
  vcfRobj = NULL,
  vcfploid = 2,
  PLAF = NULL,
  m_max = 5,
  k_max = 10,
  rho = 1e-05,
  e1 = 0.05,
  e2 = 0.05,
  burnin = 100,
  samples = 1000,
  verbose = TRUE,
  reportIteration = 1000,
  parallelize = TRUE
)
}
\description{
Run the HMMERTIME model within a Bayesian framework, using Markov Chain Monte Carlo to sample the posterior
}
